refactor: NavItem 파일에 맞게 폴더 이름 수정 (#438)
