--- frontend/src/components/@common/NavButton/index.tsx
@@ -1,3 +1,0 @@
-import NavButton from '~/components/@common/NavButton/NavButton';
-
-export default NavButton;
