+++ frontend/src/components/@common/NavItem/index.tsx
@@ -1,0 +1,3 @@
+import NavItem from '~/components/@common/NavItem/NavItem';
+
+export default NavItem;
